requirements.txt
# Core Flask
Flask==2.3.3
gunicorn==21.2.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5

# Authentication & Security
Flask-Login==0.6.3
Werkzeug==2.3.7

# Forms
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.0

# Google Generative AI
google-generativeai==0.3.1

# Environment variables
python-dotenv==1.0.0

# CSV and Data Processing
pandas==2.1.1

# Date/Time
python-dateutil==2.8.2
